refactor: move transaction Event to miden-lib
